refactor: ec2-instance-create-constants 위치, 파일명, 경로 수정
