+++ apps/web/src/components/aws-services/registry/registry.ts
@@ -10,12 +10,12 @@ import S3BucketCreate from '@/components/aws-services/s3/s3-bucket-create/s3-buc
 import S3BucketDetail from '@/components/aws-services/s3/s3-bucket-detail/s3-bucket-detail'
 import S3BucketList from '@/components/aws-services/s3/s3-bucket-list/s3-bucket-list'
 import S3FileUpload from '@/components/aws-services/s3/s3-file-upload/s3-file-upload'
-import { EC2_INSTANCE_CREATE_SECTIONS } from '@/constants/aws-services/ec2'
 import { CLOUDFRONT_CACHE_BEHAVIOR_SECTIONS } from '@/types/aws-services/cloudfront/cache-behavior/constants'
 import { CLOUDFRONT_DISTRIBUTION_LIST_SECTIONS } from '@/types/aws-services/cloudfront/distribution-list/constants'
 import { CLOUDFRONT_DISTRIBUTION_SETTINGS_SECTIONS } from '@/types/aws-services/cloudfront/distribution-settings/constants'
 import { CLOUDFRONT_ORIGIN_SETTINGS_SECTIONS } from '@/types/aws-services/cloudfront/origin-settings/constants'
 import { CLOUDFRONT_WEBSITE_SETTINGS_SECTIONS } from '@/types/aws-services/cloudfront/website-settings/constants'
+import { EC2_INSTANCE_CREATE_SECTIONS } from '@/types/aws-services/ec2/instance-create/constants'
 import { S3_BUCKET_CREATE_SECTIONS } from '@/types/aws-services/s3/bucket-create/'
 import { S3_BUCKET_DETAIL_SECTIONS } from '@/types/aws-services/s3/bucket-detail/'
 import { S3_BUCKET_LIST_SECTIONS } from '@/types/aws-services/s3/bucket-list/'
