+++ apps/web/src/constants/aws-services/ec2/index.ts
@@ -6,8 +6,3 @@ export {
   FIREWALL_OPTIONS,
 } from './ec2-options.constants'
 export { EC2_TOOLTIPS } from './ec2-tooltips.constants'
-export {
-  EC2_INSTANCE_CREATE_SECTIONS,
-  type EC2InstanceCreateSectionKey,
-  type EC2InstanceCreateConfig,
-} from './ec2-instance-create.constants'
