+++ apps/web/src/types/aws-services/ec2/instance-create/constants.ts
