+++ apps/web/src/types/aws-services/ec2/instance-create/index.ts
@@ -3,3 +3,8 @@ export type {
   EC2SectionProps,
   EC2WithSetValueSectionProps,
 } from './ec2-config.types'
+export {
+  EC2_INSTANCE_CREATE_SECTIONS,
+  type EC2InstanceCreateSectionKey,
+  type EC2InstanceCreateConfig,
+} from './constants'
